setelah sekian lama

+++ database/migrations/2024_10_17_061302_create_model_detail_pribadis_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('detail_pribadi', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('id_user');
-            $table->integer('id_maskapai');
+            $table->unsignedBigInteger('id_maskapai');
             $table->string('name');
             $table->string('email');
             $table->string('phone');
