+++ database/migrations/2024_10_17_061310_create_model_detail_pembayarans_table.php
@@ -14,7 +14,7 @@ return new class extends Migration
         Schema::create('detail_pembayaran', function (Blueprint $table) {
             $table->id();
             $table->unsignedBigInteger('id_user');
-            $table->integer('id_maskapai');
+            $table->unsignedBigInteger('id_maskapai');
             $table->string('origin');
             $table->integer('penumpang_dewasa');
             $table->integer('penumpang_anak');
